Downgrade our build machine to use 13.2
This should address an issue where multiple users are seeing crashes
when trying to invoke some method on Set. This looks related to this
issue: https://bugs.swift.org/browse/SR-16096

The linked issue indicates that this is an issue with builds compiled
against Swift 5.6 but linked against Swift 5.5 at runtime. To fix this,
we should revert to building Signal with Swift 5.5 until Xcode includes
the Swift 5.6.1 patch.