Fix bug that asks users to upgrade prematurely
As of 9662b3cb1e7b47, the app should wait for a week before notifying
users about a new version. However, the “has it been 7 days” marker
isn’t cleared when a user does upgrade, so the prompt can appear too
quickly when the next update is released.

For example:
- Assume it’s Apr 7 and a user is running v5.34.
- On Apr 8, v5.35 is released & the user sets the 7-day marker to Apr 8.
- On Apr 10, the user upgrades to v5.35; the marker isn’t cleared.
- On Apr 15, v5.36 is released.
- On Apr 16, the user launches the app and sees an update banner, even
  though the new version has been out for only 1 day.