Use correct currency in UI when starting a subscription
I introduced a bug in 51247f7d74040540e187bbcad10955a07e50688e. When
starting a new subscription, changing the currency would not update the
amount in the UI. (It was correct once you got to the Apple Pay step,
though.)

For example, if you changed from USD to GBP, the UI would still show
"$5/month" instead of "£4/month".

This was because I used the wrong value, which should now be fixed.

This bug was reported on [the beta thread for 5.37][0].

[0]: https://community.signalusers.org/t/beta-feedback-for-the-upcoming-ios-5-37-release/44054