Don’t allow migration blocks to throw
In a01ad40f2236b7, migration blocks were wrapped in `do {} catch {}`
blocks so that crash stack traces would identify the specific migration
that encountered an error.

This commit enforces the recommendation from [this comment][1].

It also updates a few migrations that weren’t following the
recommendation. Errors in these migrations still result in a [crash][2],
but the stack trace doesn’t include the specific migration that failed.

[1]: https://github.com/signalapp/Signal-iOS/blob/873ba1e511b4722f9fc9db03b29902153f944929/SignalServiceKit/src/Storage/Database/GRDBSchemaMigrator.swift#L241-L242
[2]: https://github.com/signalapp/Signal-iOS/blob/873ba1e511b4722f9fc9db03b29902153f944929/SignalServiceKit/src/Storage/Database/GRDBSchemaMigrator.swift#L25