Move emoji availability cache out of the database
Emoji availability is tested by *drawing the emoji to a graphics
context* and seeing if it comes out as a replacement character glyph.
This is not the fastest operation, so we cache it, and update the
cache whenever the user updates their version of iOS. This cache
previously lived in SDSKeyValueStore, i.e. in a shared table in the
database. Reading it took longer than we might like even before we get
to updating it, and it's not even secret data at all.

This commit replaces the SDSKeyValueStore with a plist file and adds a
migration to remove the SDSKeyValueStore data.