Fix some codegen designated initializers
* Ignore `exclusiveProcessIdentifier` parameter

This was added in fee1bbddad2980, but the `NS_UNAVAILABLE` declarations
weren’t updated, which led to warnings in Xcode.

* Ignore `TSThread` initializer in subclasses

* Fix `TSAttachmentPointer` ignored initializer

* Add missing initializers to `TSMessage` subclasses

This only handles the immediate subclasses. There’s still more warnings
in the extended hierarchy, but those don’t seem to have been handled as
consistently in the first place, so updating them is more involved.