Remove useless property from interactive sheet
`allowsInteractiveDismisssal` was always set to `true` and never
changed. (It also has a typo in it.) This completely removes the
variable, as it was unused.

According to `git log -S allowsInteractiveDismisssal`, it looks like
this was introduced in 9fceb91d53d89500d3d9e5d295ce7d9f338dcd63. I
couldn't find a time when it was set to `false`.