Add missing PluralAware.stringsdict files
If we have a Localizable.strings file for a specific language, we should
also include the PluralAware.stringsdict file, even if less than 80% of
the plural strings have been translated.

When adopting a brand new language, it makes sense to wait for some
threshold of strings across the entire app to be available.

However, the same doesn’t apply for the new PluralAware files. If we
already have some translations for a language, we should include all the
translations that are available.

Assume that `L.lproj/Localizable.strings` already exists. Then:

(1) If `L.lproj/PluralAware.stringsdict` isn’t included, all plurals
will appear in the fallback language.

(2) If `L.lproj/PluralAware.stringsdict` is included but there aren’t
any translations, all plurals will appear in the fallback language.

(3) If `L.lproj/PluralAware.stringsdict` is included and has even a
single translation, that string will shown in the app.

(4) If `L.lproj/PluralAware.stringsdict` is included and has 80% of its
translations, all of those translations will be shown in the app.

This change address (3) without changing the behavior for (1) and (2).