Show additional details for subscription charge failures
In c5bdf6c094ff78637440e185ceff08c796aeb966, we started to show errors
when a subscription failed to renew because of a charge failure.

Now, we show additional text depending on what the charge failure was.
For example, if you had an invalid card number, we show a special string
for that.