Update "pending changes action sheet" localization metadata
_I recommend reviewing this with whitespace changes disabled._

This is a minor change; it have no direct user impact.

We have a "pending changes action sheet" which shows up to warn users
that their changes may not be saved. The localization key and comment
indicated that its strings were only for groups, but that's not
true—it's used for lots of things, such as editing your profile or
updating a group's metadata.