Fix crash when opening wallpaper/chat color view
_I recommend reviewing this with whitespace changes disabled._

Previously, Settings → Appearance → Chat Color & Wallpaper → Chat Color
would cause a crash. This was because we had a re-entrant database
transaction.

This fixes that by removing the re-entrant transaction—a fairly simple
fix.

There was also a function here that was immediately called and not used
elsewhere, so I removed the function and inlined its contents. Not
strictly related to the crash. I believe it was originally added in
0574d23b6f06b4ccd8e9e6d6972e1151b2188219.