Check whether recipient is blocked before adding gift badge send job
If someone makes it to the confirmation screen with a blocked contact,
we should try to stop them. Specifically, we check for blocked state (1)
when the Apple Pay button is pressed (2) right before we add a gift
badge job.

Note that [the gift badge job already checks this][0], too—this is just
helpful for the UI.

Tested this by trying to send a gift to a blocked contact, and also
trying the "happy path" of sending a gift to someone I hadn't blocked.
Both worked as I expected.

[0]: https://github.com/signalapp/Signal-iOS/blob/2a0ae59ce8b58e55995508d4559e133cabe5372b/SignalMessaging/Donations/SendGiftBadgeJobQueue.swift#L163-L165