Confirm any safety number changes before sending gift badge
If someone's safety number changes before sending them a badge, we want
to confirm with the user.

We do this in two places:

1. After clicking the Apple Pay button, but before actually requesting
   Apple Pay. If there is a safety number change, this is likely to be
   the spot where we discover it.

   If the user clicks Cancel, the whole thing is stopped. If the user
   continues, we request Apple Pay.
2. Right before adding the job. If the safety number changed after the
   previous step (which is possible), we want to check right before we
   try to send something.

   Again, if the user clicks Cancel, we stop before ever charging the
   card (though we have made a couple of non-charging network calls).

The first one was pretty easy—just throw a "user canceled" error.

The second one required some minor reworking of how the Apple Pay sheet
is dismissed. Under normal conditions, we only dismiss the sheet after
the payment method is charged. But if there's a safety number change, we
need to dismiss the sheet earlier, so we can confirm with the user.