Add "urgent" boolean to outgoing messages
To improve message reliability, we're adding an urgency flag to outgoing
messages. For example, outgoing calls are urgent, but delivery receipts
are not. [Android][] and [Desktop][] have already done this work.

At a high level, I added an `isUrgent` property to `TSOutgoingMessage`s.
It defaults to `true`. Some subclasses, like
`OWSReceiptsForSenderMessage`, override it to return `false`.

Builds off of a few other commits (not necessary to understand this
commit, but might be useful for posterity):

- e858a0d9167e3eb4e42870e6688cfd61383d46cd
- 8c6d2ebe8cecc959a5837669fcab10c53036a0de
- 402b117221eda2c61b446fe58e6bc704f9543b1c
- 2ba0cd764ded2b9f74e48d447803ce84dc5234f0
- 266a4663e913e1ae1af840cee95807ab595801e6
- 8e5009bbf7a580842ea1817049de3c3040ed2e0e

[Android]: https://github.com/signalapp/Signal-Android/commit/dc04c8ed98eb565b9c3e43522a42d3463c2634d2
[Desktop]: https://github.com/signalapp/Signal-Desktop/commit/06190b14344c4f3cb164632b35a07af3d82b5b37