Remove unused application state change notification
This change should have no obvious user impact.

When the application's state changes, we fire off a notification. But
we don't actually listen to that notification, as far as I can tell—it
looks like the last listener was removed in 2019 (in
680b844f3c077ec4e762a3210195ee5573ea44e1).

This removes that notification.

`git grep -i ReportedApplicationStateDidChange` yields no results after
this change.