Deselect the selected row if a thread cannot be found
This is required for the case where the DetailVC is presenting an
archived conversation, but the ConversationListVC is in inbox mode. In
this case, the last selected thread would continue to remain selected.

Ideally, the ConversationListVC would be smart enough to know whether or
not a thread is archived or not and transition itself accordingly.
That'd be a much larger change. For now, we select the thread if we can
and just deselect otherwise.