Repurpose -[ChatListVC lastViewedThread]
-lastViewedThread appears to be set on thread presentation and accessed
whenever the ChatList reappears to ensure that it's visible. That's
pretty similar to ensuring selection, so I've combined both.

Now, instead of exposing a property a method is exposed:
-updateLastViewedThread:animated:

The exact same behvior of adjusting the scrollOffset on appear is
maintained. In addition, the moment the lastViewedThread is updated, the
ChatList will ensure that the lastViewedThread is also selected.